requirements.txt
# API
Flask==3.1.3
flask-cors==6.0.2
python-dotenv==1.0.1

# Web Search
tavily-python==0.7.12
requests==2.32.5

# Data Processing
pandas==3.0.1
pydantic==2.9.2